+++ packages/splat-transform-native/source/include/node/api_splat.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <napi.h>
 
-namespace node_api::gaussian {
+namespace node_api::splat {
 Napi::Value generate_lod(const Napi::CallbackInfo& info);
-} // namespace node_api::gaussian
+} // namespace node_api::splat