+++ packages/splat-transform-native/source/include/splat/splat_lod.h
@@ -1,31 +1,29 @@
 #pragma once
 #include <container_helpers.h>
 #include <cstdint>
-#include <gaussian/gaussian.h>
+#include <splat/splat.h>
 #include <vector>
 
-namespace gaussian::lod {
-struct GaussianLodParameters;
-
+namespace splat::lod {
 namespace detail {
 Splat reduce_gaussians(size_t id, const Splat& splat, size_t target_count, float scale_boost, size_t max_step);
 }
 
-struct GaussianLod {
+struct SplatLod {
     std::vector<Splat> splats;
     std::vector<uint32_t> levels;
 };
 
-struct GaussianLevelParameters {
+struct SplatLevelParameters {
     float precision;
     float scale_boost;
 };
 
 template<typename T>
     requires requires(T) {
-        requires helpers::container::range_of<T, GaussianLevelParameters>;
+        requires helpers::container::range_of<T, SplatLevelParameters>;
     }
-inline void generate_lod(size_t id, GaussianLod& lod, T&& parameters, size_t min_size, size_t max_step) {
+inline void generate_lod(size_t id, SplatLod& lod, T&& parameters, size_t min_size, size_t max_step) {
     auto base_count = lod.splats.back().gaussians.size();
     for (auto& parameter : parameters) {
         auto target_count = std::max<size_t>(static_cast<size_t>(std::floor(static_cast<double>(base_count) * parameter.precision)), 1);
@@ -42,4 +40,4 @@ inline void generate_lod(size_t id, GaussianLod& lod, T&& parameters, size_t min
         lod.levels.push_back(lod.splats.size() - 1);
     }
 }
-} // namespace gaussian::lod
+} // namespace splat::lod